#161638906 Code refactor

+++ app/api/V2/views/products.py
@@ -1,24 +1,27 @@
 from flask_restful import Resource, reqparse
 from flask import request, json, jsonify, make_response
-from flask_jwt_extended import jwt_required
-from app.api.V2.models import ProductModel
-from app.api.V2.views.users import admin_only
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from functools import wraps
 
+from app.api.V2.models import ProductModel, UserModel
 
-class Products(Resource, ProductModel):
-    parser = reqparse.RequestParser()
-    parser.add_argument('name', required=True, help='Product name cannot be blank', type=str)
-    parser.add_argument('price', required=True, help=' Product price cannot be blank or a word', type=int)
-    parser.add_argument('category', required=True, help="Category must be specified", type=str)
-    parser.add_argument('available_stock', required=True, help="Define available stock", type=int)
-    parser.add_argument('min_stock', required=True, help="Define minimum stock", type=int)
+
+parser = reqparse.RequestParser()
+parser.add_argument('name', required=True, help='Product name cannot be blank', type=str)
+parser.add_argument('price', required=True, help=' Product price cannot be blank or a word', type=int)
+parser.add_argument('category', required=True, help="Category must be specified", type=str)
+parser.add_argument('inventory', required=True, help="Define available stock", type=int)
+parser.add_argument('minimum_stock', required=True, help="Define minimum stock", type=int)
+
+
+class Products(Resource, ProductModel, UserModel):
 
     def __init__(self):
         self.operation = ProductModel()
 
     @jwt_required
     def get(self):
+
         products = self.operation.get_all_products()
         if not products:
             return {"message": "No products yet"}
@@ -29,36 +32,45 @@ class Products(Resource, ProductModel):
 
     @jwt_required
     def post(self):
-        args = Products.parser.parse_args()
+        args = parser.parse_args()
         name = args.get('name').strip()  # removes whitespace
         price = args.get('price')
         category = args.get('category')
-        available_stock = args.get('available_stock')
-        min_stock = args.get('min_stock')
-
+        inventory = args.get('inventory')
+        minimum_stock = args.get('minimum_stock')
         try:
-            product = self.operation.get_item_if_exists(name, price, available_stock, min_stock, category)
+            user = UserModel.find_by_email(get_jwt_identity())
+            if user[4] != "admin":
+                return {"message": "You do not have authorization to access this feature"}
+            product = self.operation.get_item_if_exists(name, price, inventory, minimum_stock, category)
             return product
         except Exception as my_exception:
             print(my_exception)
             return {'message': 'Something went wrong.'}, 500
 
 
-class SingleProduct(Resource, ProductModel):
+class SingleProduct(Resource, ProductModel, UserModel):
 
-    @jwt_required
     def get(self, id):
         return ProductModel.get_each_product(self, id)
 
     @jwt_required
     def delete(self, id):
+        user = UserModel.find_by_email(get_jwt_identity())
+        if user[4] != "admin":
+            return {"message": "You do not have authorization to access this feature"}
         return ProductModel.delete_product(self, id)
 
     @jwt_required
     def put(self, id):
-        data = request.get_json()
-        price = data['price']
-        available_stock = data['available_stock']
-        category = data['category']
-        min_stock = data['min_stock']
-        return ProductModel.update_product(self, id, price, available_stock, min_stock, category)
+        args = parser.parse_args()
+        name = args.get('name')
+        price = args.get('price')
+        category = args.get('category')
+        inventory = args.get('inventory')
+        minimum_stock = args.get('minimum_stock')
+
+        user = UserModel.find_by_email(get_jwt_identity())
+        if user[4] != "admin":
+            return {"message": "You do not have authorization to access this feature"}
+        return ProductModel.update_product(self, id, price, inventory, minimum_stock, category)
